Package directory refactoring (changing of directories naming mostly). All imports statements changed accordingly.
