+++ esm/base/model.py
@@ -1,12 +1,12 @@
 from pathlib import Path
 from typing import Any
 
-from src import constants
-from src.log_exc.logger import Logger
-from src.log_exc.exceptions import *
-from src.support.file_manager import FileManager
-from src.support.pbi_manager import PBIManager
-from src.backend.core import Core
+from esm import constants
+from esm.log_exc.logger import Logger
+from esm.log_exc.exceptions import *
+from esm.support.file_manager import FileManager
+from esm.support.pbi_manager import PBIManager
+from esm.base.core import Core
 
 
 class Model:
