+++ esm/base/problem.py
@@ -5,13 +5,13 @@ import pandas as pd
 import numpy as np
 import cvxpy as cp
 
-from src import constants
-from src.log_exc import exceptions as exc
-from src.log_exc.logger import Logger
-from src.support import util
-from src.support.file_manager import FileManager
-from src.support.dotdict import DotDict
-from src.backend.index import Index, Variable
+from esm import constants
+from esm.log_exc import exceptions as exc
+from esm.log_exc.logger import Logger
+from esm.support import util
+from esm.support.file_manager import FileManager
+from esm.support.dotdict import DotDict
+from esm.base.index import Index, Variable
 
 
 class Problem:
