+++ esm/base/set.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict, Iterator, Tuple
 import pandas as pd
 
-from src.log_exc.logger import Logger
+from esm.log_exc.logger import Logger
 
 
 class Set:
