+++ esm/base/variable.py
@@ -2,10 +2,10 @@ from typing import Any, Dict, Iterator, List, Literal, Tuple
 import numpy as np
 
 import pandas as pd
-from src import constants
-from src.log_exc import exceptions as exc
-from src.log_exc.logger import Logger
-from src.support import util
+from esm import constants
+from esm.log_exc import exceptions as exc
+from esm.log_exc.logger import Logger
+from esm.support import util
 
 
 class Variable:
