+++ tests/test_util.py
@@ -1,5 +1,5 @@
 import pytest
-from src.support.util import *
+from esm.support.util import *
 
 
 def test_validate_selection():
